--- apps/api/src/donations/events/payment-created.service.ts
@@ -1,53 +1,0 @@
-import Stripe from 'stripe'
-import { BadRequestException, Injectable, Logger } from '@nestjs/common'
-import { StripeWebhookHandler } from '@golevelup/nestjs-stripe'
-
-import { DonationMetadata } from '../dontation-metadata.interface'
-import { CampaignService } from '../../campaign/campaign.service'
-import { getPaymentData } from '../helpers/payment-intent-helpers'
-import { DonationStatus } from '@prisma/client'
-
-@Injectable()
-export class PaymentCreatedService {
-  constructor(private campaignService: CampaignService) {}
-
-  @StripeWebhookHandler('payment_intent.created')
-  async handlePaymentIntentCreated(event: Stripe.Event) {
-    const paymentIntent: Stripe.PaymentIntent = event.data.object as Stripe.PaymentIntent
-
-    Logger.debug(
-      '[ handlePaymentIntentCreated ]',
-      paymentIntent,
-      paymentIntent.metadata as DonationMetadata,
-    )
-
-    const metadata: DonationMetadata = paymentIntent.metadata as DonationMetadata
-    if (!metadata.campaignId) {
-      throw new BadRequestException(
-        'Payment intent metadata does not contain target campaignId. Probably wrong session initiation. Payment intent is:  ' +
-          paymentIntent.id,
-      )
-    }
-
-    const campaign = await this.campaignService.getCampaignById(metadata.campaignId)
-
-    if (campaign.currency !== paymentIntent.currency.toUpperCase()) {
-      throw new BadRequestException(
-        `Donation in different currency is not allowed. Campaign currency ${
-          campaign.currency
-        } <> donation currency ${paymentIntent.currency.toUpperCase()}`,
-      )
-    }
-
-    const billingDetails = getPaymentData(paymentIntent)
-
-    /*
-     * Handle the create event
-     */
-    await this.campaignService.updateDonationPayment(
-      campaign,
-      billingDetails,
-      DonationStatus.waiting,
-    )
-  }
-}
