+++ apps/api/src/donations/events/stripe-payment.service.ts
@@ -1,0 +1,115 @@
+import Stripe from 'stripe'
+import { BadRequestException, Injectable, Logger } from '@nestjs/common'
+import { StripeWebhookHandler } from '@golevelup/nestjs-stripe'
+
+import { DonationMetadata } from '../dontation-metadata.interface'
+import { CampaignService } from '../../campaign/campaign.service'
+import { getPaymentData } from '../helpers/payment-intent-helpers'
+import { DonationStatus } from '@prisma/client'
+
+/** Testing Stripe on localhost is described here: 
+ * https://github.com/podkrepi-bg/api/blob/master/TESTING.md#testing-stripe
+*/
+@Injectable()
+export class StripePaymentService {
+  constructor(private campaignService: CampaignService) {}
+
+  @StripeWebhookHandler('payment_intent.created')
+  async handlePaymentIntentCreated(event: Stripe.Event) {
+    const paymentIntent: Stripe.PaymentIntent = event.data.object as Stripe.PaymentIntent
+
+    Logger.debug(
+      '[ handlePaymentIntentCreated ]',
+      paymentIntent,
+      paymentIntent.metadata as DonationMetadata,
+    )
+
+    const metadata: DonationMetadata = paymentIntent.metadata as DonationMetadata
+    if (!metadata.campaignId) {
+      throw new BadRequestException(
+        'Payment intent metadata does not contain target campaignId. Probably wrong session initiation. Payment intent is:  ' +
+          paymentIntent.id,
+      )
+    }
+
+    const campaign = await this.campaignService.getCampaignById(metadata.campaignId)
+
+    if (campaign.currency !== paymentIntent.currency.toUpperCase()) {
+      throw new BadRequestException(
+        `Donation in different currency is not allowed. Campaign currency ${
+          campaign.currency
+        } <> donation currency ${paymentIntent.currency.toUpperCase()}`,
+      )
+    }
+
+    const billingDetails = getPaymentData(paymentIntent)
+
+    /*
+     * Handle the create event
+     */
+    await this.campaignService.updateDonationPayment(
+      campaign,
+      billingDetails,
+      DonationStatus.waiting,
+    )
+  }
+
+  @StripeWebhookHandler('payment_intent.canceled')
+  async handlePaymentIntentCancelled(event: Stripe.Event) {
+    const paymentIntent: Stripe.PaymentIntent = event.data.object as Stripe.PaymentIntent
+    Logger.log(
+      '[ handlePaymentIntentCancelled ]',
+      paymentIntent,
+      paymentIntent.metadata as DonationMetadata,
+    )
+
+    const metadata: DonationMetadata = paymentIntent.metadata as DonationMetadata
+    if (!metadata.campaignId) {
+      throw new BadRequestException(
+        'Payment intent metadata does not contain target campaignId. Probably wrong session initiation. Payment intent is:  ' +
+          paymentIntent.id,
+      )
+    }
+
+    const campaign = await this.campaignService.getCampaignById(metadata.campaignId)
+
+    const billingData = getPaymentData(paymentIntent)
+    await this.campaignService.updateDonationPayment(
+      campaign,
+      billingData,
+      DonationStatus.cancelled,
+    )
+  }
+
+  @StripeWebhookHandler('payment_intent.succeeded')
+  async handlePaymentIntentSucceeded(event: Stripe.Event) {
+    const paymentIntent: Stripe.PaymentIntent = event.data.object as Stripe.PaymentIntent
+    Logger.log(
+      '[ handlePaymentIntentSucceeded ]',
+      paymentIntent,
+      paymentIntent.metadata as DonationMetadata,
+    )
+
+    const metadata: DonationMetadata = paymentIntent.metadata as DonationMetadata
+    if (!metadata.campaignId) {
+      throw new BadRequestException(
+        'Payment intent metadata does not contain target campaignId. Probably wrong session initiation. Payment intent is:  ' +
+          paymentIntent.id,
+      )
+    }
+
+    const campaign = await this.campaignService.getCampaignById(metadata.campaignId)
+
+    if (campaign.currency !== paymentIntent.currency.toUpperCase()) {
+      throw new BadRequestException(
+        `Donation in different currency is not allowed. Campaign currency ${
+          campaign.currency
+        } <> donation currency ${paymentIntent.currency.toUpperCase()}`,
+      )
+    }
+
+    const billingData = getPaymentData(paymentIntent)
+
+    await this.campaignService.donateToCampaign(campaign, billingData)
+  }
+}
